feat(arm): USART3 on the VCP, and real newlib syscalls
Implements mcu::Usart against the F7's USART registers and wires Uart1() to
USART3 on PD8/PD9 AF7, which the Nucleo routes to the ST-LINK virtual COM
port. Blocking Send and Receive poll ISR; SetRxHandler goes through the RXNE
interrupt into the same handler-table pattern exti.cpp uses. UnimplementedUart
is deleted.

Alternate-function mode is not expressible in mcu::PinDirection, and should
not be -- it is a peripheral-driver concern, not an application one. Pin
configuration now goes through ConfigurePin(), which writes OSPEEDR, PUPDR,
OTYPER and AFR before MODER, so a pin never spends a cycle driving through
whichever peripheral AF0 happens to be. GpioPin::Configure delegates to it
rather than keeping a second MODER write.

nosys.specs is gone; the board provides its own syscalls. _write reaches the
console USART with CRLF expansion, and _sbrk is bounded by the linker
script's __heap_limit, so exhausting the heap returns ENOMEM instead of
handing out addresses the stack is about to grow into.

The one non-obvious piece: newlib-nano ships __malloc_lock and
__malloc_unlock as `bx lr`. That is correct for a single-threaded program and
wrong here, because uart_echo's receive handler builds a std::vector -- it
allocates in interrupt context. A USART interrupt arriving while the main
context was inside malloc would re-enter the allocator and corrupt its arena,
surfacing as a fault somewhere unrelated much later. Both are now overridden
to mask interrupts, counting nesting and restoring the previous PRIMASK
rather than assuming it was clear.

Two details worth recording. Word length counts the parity bit, so 8 data
bits with parity needs a 9-bit word; Init rejects 9-plus-parity, which the
peripheral cannot express. Send waits for TC, not just TXE, so returning and
immediately reconfiguring cannot truncate the last character.

USART3 is a CMSIS macro, so nothing in usart.hpp names a vendor type -- the
instance is selected by mcu::UsartId. Same rule that gpio_port.hpp records.

Hardware verification pending: builds and image checks only.

+++ src/libs/board/stm32f767zi_nucleo/pin_map.hpp
@@ -30,11 +30,11 @@ constexpr PinLocation kUserButton1{mcu::GpioPort::kC, 13};
 // output shows up on /dev/ttyACM0. Alternate function 7.
 constexpr PinLocation kUart1Tx{mcu::GpioPort::kD, 8};
 constexpr PinLocation kUart1Rx{mcu::GpioPort::kD, 9};
-constexpr std::uint32_t kUart1AlternateFunction = 7;
+constexpr std::uint8_t kUart1AlternateFunction = 7;
 
 // I2C1 on the Arduino connector: D15 (SCL) and D14 (SDA). Alternate function 4.
 constexpr PinLocation kI2C1Scl{mcu::GpioPort::kB, 8};
 constexpr PinLocation kI2C1Sda{mcu::GpioPort::kB, 9};
-constexpr std::uint32_t kI2C1AlternateFunction = 4;
+constexpr std::uint8_t kI2C1AlternateFunction = 4;
 
 }  // namespace board::pin_map
